RD-3241 Added support for in-line edit of page menu items names (#1637)
