+++ app/actions/types.ts
@@ -34,7 +34,7 @@ export const PAGE_MANAGEMENT_SET_EDIT_MODE = 'PAGE_MANAGEMENT_SET_EDIT_MODE';
 
 export const SET_CONTEXT_VALUE = 'SET_CONTEXT_VALUE';
 export const CLEAR_CONTEXT = 'CLEAR_CONTEXT';
-export const RENAME_PAGE = 'RENAME_PAGE';
+export const RENAME_PAGE_MENU_ITEM = 'RENAME_PAGE_MENU_ITEM';
 export const REMOVE_WIDGET = 'REMOVE_WIDGET';
 export const MINIMIZE_WIDGETS = 'MINIMIZE_WIDGETS';
 export const MINIMIZE_TAB_WIDGETS = 'MINIMIZE_TAB_WIDGETS';
