+++ app/components/shared/PageFilter.tsx
@@ -4,7 +4,7 @@ import React, { FunctionComponent, useState } from 'react';
 import { useSelector } from 'react-redux';
 import { DropdownProps } from 'semantic-ui-react';
 import type { PageDefinition } from '../../actions/page';
-import { createPagesMap } from '../../actions/page';
+import { createPagesMap } from '../../actions/pageMenu';
 import type { ReduxState } from '../../reducers';
 
 import { Form } from '../basic';
