+++ app/components/sidebar/AddPageButton.tsx
@@ -1,29 +1,21 @@
-// @ts-nocheck File not migrated fully to TS
-/**
- * Created by kinneretzin on 29/08/2016.
- */
-
-import PropTypes from 'prop-types';
 import React from 'react';
-import { connect } from 'react-redux';
+import { useDispatch } from 'react-redux';
 import i18n from 'i18next';
-import { addPage } from '../../actions/page';
+import { addPage } from '../../actions/pageMenu';
 import EditModeButton from '../EditModeButton';
 
 let nameIndex = 0;
 
-const mapDispatchToProps = dispatch => {
-    return {
-        onClick: () => {
-            dispatch(addPage(i18n.t('editMode.defaultPageName', 'Page_{{index}}', { index: nameIndex })));
-            nameIndex += 1;
-        }
-    };
-};
-
 export const buttonWidth = 146;
 
-const AddPageButton = ({ onClick }) => {
+const AddPageButton = () => {
+    const dispatch = useDispatch();
+
+    function onClick() {
+        dispatch(addPage(i18n.t('editMode.defaultPageName', 'Page_{{index}}', { index: nameIndex })));
+        nameIndex += 1;
+    }
+
     return (
         <EditModeButton
             icon="add"
@@ -36,10 +28,4 @@ const AddPageButton = ({ onClick }) => {
     );
 };
 
-AddPageButton.propTypes = {
-    onClick: PropTypes.func.isRequired
-};
-
-const AddPage = connect(null, mapDispatchToProps)(AddPageButton);
-
-export default AddPage;
+export default AddPageButton;
