+++ app/components/sidebar/PagesList.tsx
@@ -1,4 +1,3 @@
-// @ts-nocheck File not migrated fully to TS
 import type { FunctionComponent } from 'react';
 import React, { ReactNode, useCallback, useMemo, useState } from 'react';
 import { chain, find, includes, map, without } from 'lodash';
@@ -6,46 +5,79 @@ import type { DragEndEvent, DragOverEvent, DragStartEvent } from '@dnd-kit/core'
 import { closestCenter, DndContext, PointerSensor, useSensor, useSensors } from '@dnd-kit/core';
 import { SortableContext, verticalListSortingStrategy } from '@dnd-kit/sortable';
 import { restrictToVerticalAxis } from '@dnd-kit/modifiers';
-import { connect, useDispatch } from 'react-redux';
+import { useDispatch, useSelector } from 'react-redux';
+import styled from 'styled-components';
 
-import { Icon } from '../basic';
+import { EditableLabel, Icon } from '../basic';
 import AddPageButton from './AddPageButton';
 import AddPageGroupButton from './AddPageGroupButton';
-import SortableMenuItem from './SortableMenuItem';
+import SortableMenuItem, { ItemContainer } from './SortableMenuItem';
 
-import type { PageDefinition, PageMenuItem } from '../../actions/page';
-import Consts from '../../utils/consts';
-import { useBoolean } from '../../utils/hooks';
-
+import type { PageDefinition } from '../../actions/page';
+import type { PageMenuItem } from '../../actions/pageMenu';
 import {
+    changePageMenuItemName,
     createPagesMap,
+    findSelectedRootPageId,
     InsertPosition,
-    removePageWithChildren,
-    removeSinglePageMenuItem,
+    removePageMenuItem,
     reorderPageMenu,
     selectPage
-} from '../../actions/page';
+} from '../../actions/pageMenu';
+import Consts from '../../utils/consts';
+import { useBoolean, useResettableState } from '../../utils/hooks';
+import { ReduxState } from '../../reducers';
 
 export interface PagesListProps {
-    onPageSelected: (page: PageDefinition) => void;
-    onItemRemoved: (page: PageMenuItem) => void;
-    pages: PageMenuItem[];
-    selected?: string;
-    isEditMode: boolean;
+    isEditMode?: boolean;
+    pageId: string;
 }
 
 function drillDownPagesFilter(pageMenuItem: PageMenuItem) {
     return pageMenuItem.type === 'pageGroup' || !pageMenuItem.isDrillDown;
 }
 
-const PagesList: FunctionComponent<PagesListProps> = ({
-    isEditMode,
-    onItemRemoved,
-    onPageSelected,
-    pages,
-    selected = ''
-}) => {
+function consumeEvent(event: React.MouseEvent) {
+    event.stopPropagation();
+    event.preventDefault();
+}
+
+const RemoveIcon = styled(Icon)`
+    position: relative;
+    top: 3px;
+    right: 3px;
+    visibility: hidden;
+    float: right;
+
+    ${ItemContainer}:hover & {
+        visibility: visible;
+    }
+`;
+
+const EditIcon = styled(Icon)`
+    position: relative;
+    top: -3px;
+    margin-left: 1em !important;
+    display: none !important;
+
+    ${ItemContainer}:hover & {
+        display: inline !important;
+    }
+`;
+
+const PagesList: FunctionComponent<PagesListProps> = ({ isEditMode, pageId }) => {
     const [expandedGroupIds, setExpandedGroupIds] = useState<string[]>([]);
+    const [dragForbidden, setDragForbidden, unsetDragForbidden] = useBoolean();
+    const [dragging, setDragging, unsetDragging] = useBoolean();
+    const [nameEditedMenuItemId, setNameEditedMenuItemId, stopNameEdit] = useResettableState<string | null>(null);
+
+    const pages = useSelector((state: ReduxState) => state.pages);
+    const selected = useMemo(() => {
+        const pagesMap = createPagesMap(pages);
+        const page = pagesMap[pageId];
+        const homePageId = pages[0].id;
+        return pages && pages.length > 0 ? findSelectedRootPageId(pagesMap, page ? page.id : homePageId) : null;
+    }, [pages, pageId]);
     const pageMenuItemsIds = useMemo(
         () =>
             chain(pages)
@@ -60,11 +92,10 @@ const PagesList: FunctionComponent<PagesListProps> = ({
         [pages, expandedGroupIds]
     );
     const pageIds = useMemo(() => pages.filter(drillDownPagesFilter).map(({ id }) => id), [pages]);
+
+    const dispatch = useDispatch();
     const sensors = useSensors(useSensor(PointerSensor, { activationConstraint: { distance: 1 } }));
     const pageCount = pageIds.length;
-    const [dragForbidden, setDragForbidden, unsetDragForbidden] = useBoolean();
-    const [dragging, setDragging, unsetDragging] = useBoolean();
-    const dispatch = useDispatch();
 
     const handleDragEnd = useCallback(
         (event: DragEndEvent) => {
@@ -141,7 +172,28 @@ const PagesList: FunctionComponent<PagesListProps> = ({
         else setExpandedGroupIds([...expandedGroupIds, clickedPageGroupId]);
     }
 
+    function onPageSelected(page: PageDefinition) {
+        dispatch(selectPage(page.id, page.isDrillDown));
+    }
+
+    function onItemRemoved(pageListItem: PageMenuItem) {
+        dispatch(removePageMenuItem(pageListItem, pageId));
+    }
+
+    function onNameChange(pageMenuItemId: string, newName: string) {
+        dispatch(changePageMenuItemName(pageMenuItemId, newName));
+        stopNameEdit();
+    }
+
     function renderPageMenuItem(pageMenuItem: PageMenuItem, subItem = false): ReactNode[] {
+        const itemNameInEdit = nameEditedMenuItemId === pageMenuItem.id;
+
+        function calculateLeftPadding() {
+            if (subItem && !itemNameInEdit) return 25;
+            if (itemNameInEdit) return 0;
+            return undefined;
+        }
+
         const renderedMenuItem = (
             <SortableMenuItem
                 id={pageMenuItem.id}
@@ -151,44 +203,67 @@ const PagesList: FunctionComponent<PagesListProps> = ({
                 href={`${Consts.CONTEXT_PATH}/page/${pageMenuItem.id}`}
                 active={selected === pageMenuItem.id}
                 className={`pageMenuItem ${pageMenuItem.id}PageMenuItem`}
-                onClick={event => {
-                    event.stopPropagation();
-                    event.preventDefault();
+                onClick={(event: React.MouseEvent) => {
+                    consumeEvent(event);
                     if (pageMenuItem.type === 'page') onPageSelected(pageMenuItem);
                     else onPageGroupClick(pageMenuItem.id);
                 }}
-                style={{ paddingLeft: subItem ? 25 : undefined, cursor: dragging ? 'inherit' : undefined }}
+                style={{
+                    height: 37,
+                    paddingTop: itemNameInEdit ? 3 : undefined,
+                    paddingLeft: calculateLeftPadding(),
+                    paddingBottom: 3,
+                    paddingRight: 0,
+                    cursor: dragging ? 'inherit' : undefined
+                }}
             >
-                {pageMenuItem.name}
-                {isEditMode && (
+                <EditableLabel
+                    value={pageMenuItem.name}
+                    editing={itemNameInEdit}
+                    onCancel={stopNameEdit}
+                    onChange={newName => onNameChange(pageMenuItem.id, newName)}
+                    style={{
+                        color: 'inherit',
+                        float: 'none',
+                        padding: 0,
+                        margin: 0,
+                        fontSize: 'inherit',
+                        fontWeight: 'inherit',
+                        ...(!itemNameInEdit && {
+                            height: 13,
+                            maxWidth: 113,
+                            overflow: 'hidden',
+                            textOverflow: 'ellipsis'
+                        })
+                    }}
+                />
+                {isEditMode && !nameEditedMenuItemId && (
                     <>
-                        <Icon
+                        <EditIcon
                             name="edit"
                             size="small"
-                            style={{ float: 'none' }}
-                            onClick={(event: MouseEvent) => {
-                                event.preventDefault();
-                                event.stopPropagation();
+                            onClick={(event: React.MouseEvent) => {
+                                consumeEvent(event);
+                                setNameEditedMenuItemId(pageMenuItem.id);
                             }}
                         />
                         {pageCount > 1 && (
-                            <Icon
+                            <RemoveIcon
                                 name="remove"
                                 size="small"
-                                className="pageRemoveButton"
-                                onClick={(event: MouseEvent) => {
-                                    event.preventDefault();
-                                    event.stopPropagation();
+                                onClick={(event: React.MouseEvent) => {
+                                    consumeEvent(event);
                                     onItemRemoved(pageMenuItem);
                                 }}
                             />
                         )}
                     </>
                 )}
-                {pageMenuItem.type === 'pageGroup' && (
+                {pageMenuItem.type === 'pageGroup' && !itemNameInEdit && (
                     <Icon
                         name="dropdown"
                         rotated={includes(expandedGroupIds, pageMenuItem.id) ? undefined : 'counterclockwise'}
+                        style={{ position: 'absolute', right: 12, margin: 0 }}
                     />
                 )}
             </SortableMenuItem>
@@ -228,7 +303,6 @@ const PagesList: FunctionComponent<PagesListProps> = ({
                     </SortableContext>
                 </DndContext>
                 <div style={{ textAlign: 'center', marginTop: 10 }}>
-                    {/* @ts-expect-error until the AddPageButton not fully migrated to ts */}
                     <AddPageButton />
                     <AddPageGroupButton />
                 </div>
@@ -239,76 +313,4 @@ const PagesList: FunctionComponent<PagesListProps> = ({
     return pagesContainer;
 };
 
-const findSelectedRootPageId = (pagesMap, selectedPageId) => {
-    const getParentPageId = page => {
-        if (!page.parent) {
-            return page.id;
-        }
-        return getParentPageId(pagesMap[page.parent]);
-    };
-
-    return getParentPageId(pagesMap[selectedPageId]);
-};
-
-const mapStateToProps = (state, ownProps) => {
-    const { pages } = state;
-    const pagesMap = createPagesMap(pages);
-    const page = pagesMap[ownProps.pageId];
-    const homePageId = pages[0].id;
-    const pageId = page ? page.id : homePageId;
-    const selected = pages && pages.length > 0 ? findSelectedRootPageId(pagesMap, pageId) : null;
-
-    return {
-        pages,
-        selected
-    };
-};
-
-function mergeProps(stateProps, dispatchProps, ownProps) {
-    return {
-        ...ownProps,
-        ...dispatchProps,
-        ...stateProps,
-        onItemRemoved: pageListItem => {
-            dispatchProps.onItemRemoved(pageListItem, stateProps.pages);
-        }
-    };
-}
-
-const mapDispatchToProps = (dispatch, ownProps) => {
-    return {
-        onPageSelected: page => {
-            dispatch(selectPage(page.id, page.isDrillDown));
-        },
-        onItemRemoved: (pageListItem, pages) => {
-            const pagesMap = createPagesMap(pages);
-            const selectedRootPageId = findSelectedRootPageId(pagesMap, ownProps.pageId);
-
-            if (pageListItem.type === 'page') {
-                // Check if user removes current page
-                if (selectedRootPageId === pageListItem.id) {
-                    // Check if current page is home page
-                    if (selectedRootPageId === ownProps.homePageId) {
-                        dispatch(selectPage(pages[1].id, false));
-                    } else {
-                        dispatch(selectPage(ownProps.homePageId, false));
-                    }
-                }
-
-                dispatch(removePageWithChildren(pageListItem));
-            } else {
-                // Check if current page is in group being removed
-                if (includes(pageListItem.pages, selectedRootPageId)) {
-                    // Select first page that is not in the group
-                    dispatch(selectPage(find(pagesMap, page => !includes(pageListItem.pages, page)).id));
-                }
-
-                dispatch(removeSinglePageMenuItem(pageListItem.id));
-            }
-        }
-    };
-};
-
-const ConnectedPagesList = connect(mapStateToProps, mapDispatchToProps, mergeProps)(PagesList);
-
-export default ConnectedPagesList;
+export default PagesList;
