+++ app/reducers/index.ts
@@ -7,7 +7,7 @@ import { combineReducers } from 'redux';
 import { connectRouter } from 'connected-react-router';
 import type { History } from 'history';
 
-import pages from './pageReducer';
+import pageMenuItemsReducer from './pageReducer';
 import widgetDefinitions from './widgetDefinitionsReducer';
 import templates from './templatesReducer';
 import templateManagement from './templateManagementReducer';
@@ -24,7 +24,7 @@ const rootReducer = (history: History) =>
         router: connectRouter(history),
         app,
         manager,
-        pages,
+        pages: pageMenuItemsReducer,
         widgetDefinitions,
         templates,
         templateManagement,
