+++ package.json
@@ -56,7 +56,7 @@
     "blueimp-md5": "^2.7.0",
     "cloudify-blueprint-topology": "^2.5.1",
     "cloudify-ui-common": "^4.1.0",
-    "cloudify-ui-components": "^2.8.1",
+    "cloudify-ui-components": "^2.9.0",
     "connected-react-router": "^6.5.2",
     "d3": "^3.5.17",
     "d3-format": "^1.3.2",
