+++ test/cypress/components/app/PagesList_spec.tsx
@@ -7,7 +7,9 @@ import { mountWithProvider } from '../utils';
 
 describe('PagesList', () => {
     it('handles default mode', () => {
-        cy.fixture('pages/pages_with_groups').then(pages => mountWithProvider(<PagesList pages={pages} />, { pages }));
+        cy.fixture('pages/pages_with_groups').then(pages =>
+            mountWithProvider(<PagesList pageId={pages[0].id} />, { pages })
+        );
 
         cy.contains('Top Level Page')
             .should('be.visible')
@@ -40,11 +42,23 @@ describe('PagesList', () => {
 
     it('handles edit mode', () => {
         cy.fixture('pages/pages_with_groups').then(pages => {
-            mountWithProvider(<PagesList isEditMode={true} pages={pages} />, { pages });
+            mountWithProvider(<PagesList isEditMode pageId={pages[0].id} />, { pages });
 
             cy.contains('Add Page').should('be.visible');
             cy.get('.icon.edit').should('have.length', pages.length);
             cy.get('.icon.remove').should('have.length', pages.length);
+
+            cy.contains('Add Page Group').click();
+            cy.contains('.item', 'Page_Group_0').find('.edit').click({ force: true });
+            cy.get('input').type('2{enter}');
+            cy.contains('.item', 'Page_Group_02').find('.remove').click({ force: true });
+            cy.contains('Page_Group_02').should('not.exist');
+
+            cy.contains('Group 1').click();
+            cy.contains('.item', 'Subpage 1').find('.edit').click({ force: true });
+            cy.get('input').type('2{enter}');
+            cy.contains('.item', 'Subpage 12').find('.remove').click({ force: true });
+            cy.contains('Subpage 12').should('not.exist');
         });
     });
 });
